requirements.txt
annotated-types==0.7.0
anyio==4.6.0
AnyQt==0.2.0
asttokens==2.4.1
attrs==24.2.0
backports.tarfile==1.2.0
baycomp==1.0.3
Bottleneck==1.4.0
catboost==1.2.7
cattrs==24.1.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
dictdiffer==0.9.0
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
fonttools==4.54.1
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.4.1
keyrings.alt==5.0.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.4
openpyxl==3.1.5
openTSNE==1.0.2
orange-canvas-core==0.2.2
orange-widget-base==4.25.0
Orange3==3.37.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-louvain==0.16
python-multipart==0.0.12
pytz==2024.2
pywin32==307
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
qasync==0.27.1
qtconsole==5.6.0
QtPy==2.4.1
requests==2.32.3
requests-cache==1.2.1
rich==13.9.2
scikit-learn==1.4.2
scipy==1.14.1
serverfiles==0.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.31.1
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
xgboost==2.1.1
xlrd==2.0.1
XlsxWriter==3.2.0
zipp==3.20.2
